每个用户都可以创建题目。

+++ SdojWeb/Controllers/QuestionController.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using System.Data.Entity;
+using System.Linq;
 using System.Threading.Tasks;
 using System.Net;
 using System.Web.Mvc;
@@ -44,7 +45,7 @@ namespace SdojWeb.Controllers
         }
 
         // GET: Questions/Create
-        [SdojAuthorize(Roles = "admin")]
+        [SdojAuthorize]
         public ActionResult Create()
         {
             return View();
@@ -53,7 +54,7 @@ namespace SdojWeb.Controllers
         // POST: Questions/Create
         // 为了防止“过多发布”攻击，请启用要绑定到的特定属性，有关 
         // 详细信息，请参阅 http://go.microsoft.com/fwlink/?LinkId=317598。
-        [HttpPost, SdojAuthorize(Roles = "admin")]
+        [HttpPost, SdojAuthorize]
         [ValidateAntiForgeryToken]
         public async Task<ActionResult> Create([Bind(Include = "Name,Description,SampleInput,SampleOutput,MemoryLimitMB,TimeLimit")] Question question)
         {
@@ -69,31 +70,43 @@ namespace SdojWeb.Controllers
         }
 
         // GET: Questions/Edit/5
-        [SdojAuthorize(Roles = "admin")]
+        [SdojAuthorize]
         public async Task<ActionResult> Edit(int? id)
         {
             if (id == null)
             {
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
             }
-            Question question = await _dbContext.Questions.FindAsync(id);
+            var question = await _dbContext.Questions.FindAsync(id);
             if (question == null)
             {
-                return HttpNotFound();
+                return RedirectToAction("Index").WithError("没找到该题目。");
             }
+
+            if (!User.IsUserOrAdmin(question.CreateUserId))
+            {
+                return RedirectToAction("Index").WithWarning("仅题目创建者才能编辑题目。");
+            }
+            
             return View(question);
         }
 
         // POST: Questions/Edit/5
         // 为了防止“过多发布”攻击，请启用要绑定到的特定属性，有关 
         // 详细信息，请参阅 http://go.microsoft.com/fwlink/?LinkId=317598。
-        [HttpPost, SdojAuthorize(Roles = "admin")]
+        [HttpPost, SdojAuthorize]
         [ValidateAntiForgeryToken]
         public async Task<ActionResult> Edit([Bind(Include = "Id,Name,Description,SampleInput,SampleOutput,MemoryLimitMB,TimeLimit")] Question question)
         {
             if (ModelState.IsValid)
             {
+                var owner = _dbContext.Questions.Where(x => x.Id == question.Id).Select(x => x.CreateUserId).FirstOrDefault();
+                if (!User.IsUserOrAdmin(owner))
+                {
+                    return RedirectToAction("Index").WithWarning("仅题目创建者才能编辑题目。");
+                }
                 question.CreateTime = DateTime.Now;
+                question.CreateUserId = User.Identity.GetIntUserId();
                 _dbContext.Entry(question).State = EntityState.Modified;
                 await _dbContext.SaveChangesAsync();
                 return RedirectToAction("Index");
@@ -102,27 +115,37 @@ namespace SdojWeb.Controllers
         }
 
         // GET: Questions/Delete/5
-        [SdojAuthorize(Roles = "admin")]
-        public async Task<ActionResult> Delete(int? id)
+        [SdojAuthorize]
+        public async Task<ActionResult> Delete(int id)
         {
-            if (id == null)
-            {
-                return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
-            }
-            Question question = await _dbContext.Questions.FindAsync(id);
+            var question = await _dbContext.Questions.FindAsync(id);
             if (question == null)
             {
-                return HttpNotFound();
+                return RedirectToAction("Index").WithError("未找到该题目。");
             }
+            if (!User.IsUserOrAdmin(question.CreateUserId))
+            {
+                return RedirectToAction("Index").WithWarning("仅题目创建者才能删除题目。");
+            }
+            
             return View(question);
         }
 
         // POST: Questions/Delete/5
-        [HttpPost, ActionName("Delete"), SdojAuthorize(Roles = "admin")]
+        [HttpPost, ActionName("Delete"), SdojAuthorize]
         [ValidateAntiForgeryToken]
         public async Task<ActionResult> DeleteConfirmed(int id)
         {
-            Question question = await _dbContext.Questions.FindAsync(id);
+            var question = await _dbContext.Questions.FindAsync(id);
+            if (question == null)
+            {
+                return RedirectToAction("Index").WithError("未找到该题目。");
+            }
+            if (!User.IsUserOrAdmin(question.CreateUserId))
+            {
+                return RedirectToAction("Index").WithWarning("仅题目创建者才能删除题目。");
+            }
+
             _dbContext.Questions.Remove(question);
             await _dbContext.SaveChangesAsync();
             return RedirectToAction("Index");
